+++ utils/FileUtils.h
@@ -6,5 +6,5 @@ namespace file_utils
 {
 	bool validDirectory(const std::string& path);
 	bool createDirectory(const std::string& path, const std::string& name);
-	int countFiles(const std::string& path);
+	__int64 countFiles(const std::string& path);
 }
